Improved installer and new systemd unit file

+++ install.sh
@@ -23,7 +23,7 @@ OSYNC_SERVICE_FILE_SYSTEMD_USER="osync-srv@.service.user"
 
 ## pmocr specfic code
 PMOCR_SERVICE_FILE_INIT="pmocr-srv"
-PMOCR_SERVICE_FILE_SYSTEMD_SYSTEM="pmocr-srv@service"
+PMOCR_SERVICE_FILE_SYSTEMD_SYSTEM="pmocr-srv.service"
 
 ## Generic code
 
@@ -120,54 +120,49 @@ if [  -f "./ssh_filter.sh" ]; then
 fi
 
 # OSYNC SPECIFIC
-if ([ -f "./$OSYNC_SERVICE_FILE_INIT" ] || [ -f "./$OSYNC_SERVICE_FILE_SYSTEMD_SYSTEM" ] ); then
-	if [ "$init" == "systemd" ]; then
-		cp "./$OSYNC_SERVICE_FILE_SYSTEMD_SYSTEM" "$SERVICE_DIR_SYSTEMD_SYSTEM" && cp "./$OSYNC_SERVICE_FILE_SYSTEMD_USER" "$SERVICE_DIR_SYSTEMD_USER/$SERVICE_FILE_SYSTEMD_SYSTEM"
-		if [ $? != 0 ]; then
-			echo "Cannot copy the systemd file to [$SERVICE_DIR_SYSTEMD_SYSTEM] or [$SERVICE_DIR_SYSTEMD_USER]."
-		else
-			echo "Created osync-srv service in [$SERVICE_DIR_SYSTEMD_SYSTEM] and [$SERVICE_DIR_SYSTEMD_USER]."
-			echo "Activate with [systemctl start osync-srv@instance.conf] where instance.conf is the name of the config file in /etc/osync."
-			echo "Enable on boot with [systemctl enable osync-srv@instance.conf]."
-			echo "In userland, active with [systemctl --user start osync-srv@instance.conf]."
-		fi
-	elif [ "$init" == "init" ]; then
-		cp "./$OSYNC_SERVICE_FILE_INIT" "$SERVICE_DIR_INIT"
-		if [ $? != 0 ]; then
-			echo "Cannot copy osync-srv to [$SERVICE_DIR_INIT]."
-		else
-			chmod 755 "$SERVICE_DIR_INIT/$OSYNC_SERVICE_FILE_INIT"
-			echo "Created osync-srv service in [$SERVICE_DIR_INIT]."
-			echo "Activate with [service $OSYNC_SERVICE_FILE_INIT start]."
-			echo "Enable on boot with [chkconfig $OSYNC_SERVICE_FILE_INIT on]."
-		fi
+if ([ "$init" == "systemd" ] && [ -f "./$OSYNC_SERVICE_FILE_SYSTEMD_SYSTEM" ]); then
+	cp "./$OSYNC_SERVICE_FILE_SYSTEMD_SYSTEM" "$SERVICE_DIR_SYSTEMD_SYSTEM" && cp "./$OSYNC_SERVICE_FILE_SYSTEMD_USER" "$SERVICE_DIR_SYSTEMD_USER/$SERVICE_FILE_SYSTEMD_SYSTEM"
+	if [ $? != 0 ]; then
+		echo "Cannot copy the systemd file to [$SERVICE_DIR_SYSTEMD_SYSTEM] or [$SERVICE_DIR_SYSTEMD_USER]."
+	else
+		echo "Created osync-srv service in [$SERVICE_DIR_SYSTEMD_SYSTEM] and [$SERVICE_DIR_SYSTEMD_USER]."
+		echo "Can be activated with [systemctl start osync-srv@instance.conf] where instance.conf is the name of the config file in /etc/osync."
+		echo "Can be enabled on boot with [systemctl enable osync-srv@instance.conf]."
+		echo "In userland, active with [systemctl --user start osync-srv@instance.conf]."
+	fi
+elif [ -f "./$OSYNC_SERVICE_FILE_INIT" ]; then
+	cp "./$OSYNC_SERVICE_FILE_INIT" "$SERVICE_DIR_INIT"
+	if [ $? != 0 ]; then
+		echo "Cannot copy osync-srv to [$SERVICE_DIR_INIT]."
+	else
+		chmod 755 "$SERVICE_DIR_INIT/$OSYNC_SERVICE_FILE_INIT"
+		echo "Created osync-srv service in [$SERVICE_DIR_INIT]."
+		echo "Can be activated with [service $OSYNC_SERVICE_FILE_INIT start]."
+		echo "Can be enabled on boot with [chkconfig $OSYNC_SERVICE_FILE_INIT on]."
 	fi
 fi
 
 # PMOCR SPECIFIC
-if ([ -f "./$PMOCR_SERVICE_FILE_INIT" ] || [ -f "./$PMOCR_SERVICE_FILE_SYSTEMD_SYSTEM" ] ); then
-	if [ "$init" == "systemd" ]; then
-		cp "./$PMOCR_SERVICE_FILE_SYSTEMD_SYSTEM" "$SERVICE_DIR_SYSTEMD_SYSTEM"
-		if [ $? != 0 ]; then
-			echo "Cannot copy the systemd file to [$SERVICE_DIR_SYSTEMD_SYSTEM] or [$SERVICE_DIR_SYSTEMD_USER]."
-		else
-			echo "Created pmocr-srv service in [$SERVICE_DIR_SYSTEMD_SYSTEM] and [$SERVICE_DIR_SYSTEMD_USER]."
-			echo "Activate with [systemctl start pmocr-srv] after configuring file options in [$BIN_DIR/$PROGRAM]."
-			echo "Enable on boot with [systemctl enable pmocr-srv]."
-		fi
-	elif [ "$init" == "init" ]; then
-		cp "./$PMOCR_SERVICE_FILE_INIT" "$SERVICE_DIR_INIT"
-		if [ $? != 0 ]; then
-			echo "Cannot copy pmoct-srv to [$SERVICE_DIR_INIT]."
-		else
-			chmod 755 "$SERVICE_DIR_INIT/$PMOCR_SERVICE_FILE_INIT"
-			echo "Created osync-srv service in [$SERVICE_DIR_INIT]."
-			echo "Activate with [service $PMOCR_SERVICE_FILE_INIT start]."
-			echo "Enable on boot with [chkconfig $PMOCR_SERVICE_FILE_INIT on]."
-		fi
+if ([ "$init" == "systemd" ] && [ -f "./$PMOCR_SERVICE_FILE_SYSTEMD_SYSTEM" ]); then
+	cp "./$PMOCR_SERVICE_FILE_SYSTEMD_SYSTEM" "$SERVICE_DIR_SYSTEMD_SYSTEM"
+	if [ $? != 0 ]; then
+		echo "Cannot copy the systemd file to [$SERVICE_DIR_SYSTEMD_SYSTEM] or [$SERVICE_DIR_SYSTEMD_USER]."
+	else
+		echo "Created pmocr-srv service in [$SERVICE_DIR_SYSTEMD_SYSTEM] and [$SERVICE_DIR_SYSTEMD_USER]."
+		echo "Can be activated with [systemctl start pmocr-srv] after configuring file options in [$BIN_DIR/$PROGRAM]."
+		echo "Can be enabled on boot with [systemctl enable pmocr-srv]."
+	fi
+elif [ -f "./$PMOCR_SERVICE_FILE_INIT" ]; then
+	cp "./$PMOCR_SERVICE_FILE_INIT" "$SERVICE_DIR_INIT"
+	if [ $? != 0 ]; then
+		echo "Cannot copy pmoct-srv to [$SERVICE_DIR_INIT]."
+	else
+		chmod 755 "$SERVICE_DIR_INIT/$PMOCR_SERVICE_FILE_INIT"
+		echo "Created osync-srv service in [$SERVICE_DIR_INIT]."
+		echo "Can be activated with [service $PMOCR_SERVICE_FILE_INIT start]."
+		echo "Can be enabled on boot with [chkconfig $PMOCR_SERVICE_FILE_INIT on]."
 	fi
 fi
-
 
 function Statistics {
